vtktounity#19 Controller Interaction often fails - fix
The problem was that on the controller releasing an object a new physics collision occurred, which added the object a second time to the collision stack. But there was only one exit, so one of the collisions remained on the stack. Solution was to only add an object to the stack if it is not already there. (An alternative would be to remove the object when it is picked up.)

Definitely seems to be working much better now.

Left the debugging code in, but commented out, as the interaction can only really be debugged in this way.

Committing straight to master as it is such a minor change.