vtktounity#15 Remove canvas from MPR - the best fix I could manage - updated MPR scripts and VR scene
Rendering the back of the MPR quad seems nigh on impossible without having the canvas to 'fudge' this. I tried flipping the MPR to create the same effect, which was possible, but even trying to minimise the updates this had a big effect on MPR performance (effectively creating a second image.) I left the code in for flipping the image in case it is helpful in the future, or someone is prepared to take the performance hit for the effect.

I also tried scaling the image to see if this would reverse it, but this did not work.

So in the end I settled for just creating another actor which uses the same image, so the texture does at least show up on both sides, but they both look the same rather than being the front/back of the same image.

Commit straight to master as this is a single commit, and I'm the only one actively using the repo.